+++ 190_export_JAQUARD_distance_artefacts.sh
@@ -83,36 +83,44 @@ for i in "${!inpth_matrix[@]}"; do
     results_dir="$trpth/Zenodo/Qiime/190_"$results_tmp"_JAQUARD_distance_artefacts"
     # echo "$results_dir"
     
-    mkdir -p "$results_dir"
-    
-    # create output filenames - pcoa
-    tmp_pcoa="${inpth_pcoa[$i]:0:-4}"
-    results_pcoa="190_"$(basename "$tmp_pcoa")".txt"
-    
-    # create output filenames - matrix
-    tmp_matx="${inpth_matrix[$i]:0:-4}"
-    results_matx="190_"$(basename "$tmp_matx")".tsv"
-    
-    printf "${bold}$(date):${normal} Exporting \"$(basename "${inpth_pcoa[$i]}")\".\n"
-    # erase possibly existing temp files
-    rm -f "$TMPDIR"ordination.txt
-    # export to temp file
-    qiime tools export \
-      --input-path  "${inpth_pcoa[$i]}" \
-      --output-path "$TMPDIR"
-    # move temp file in place
-    mv "$TMPDIR"ordination.txt "$results_dir"/"$results_pcoa"
+    if [ ! -d "$results_dir" ]; then
 
-    printf "${bold}$(date):${normal} Exporting \"$(basename "${inpth_matrix[$i]}")\".\n"
-    # erase possibly existing temp files
-    rm -f "$TMPDIR"distance-matrix.tsv 
-    # export to temp file
-    qiime tools export \
-      --input-path  "${inpth_matrix[$i]}" \
-      --output-path "$TMPDIR"
-    # move temp file in place
-    mv "$TMPDIR"distance-matrix.tsv "$results_dir"/"$results_matx"
+      mkdir -p "$results_dir"
     
+      # create output filenames - pcoa
+      tmp_pcoa="${inpth_pcoa[$i]:0:-4}"
+      results_pcoa="190_"$(basename "$tmp_pcoa")".txt"
+    
+      # create output filenames - matrix
+      tmp_matx="${inpth_matrix[$i]:0:-4}"
+      results_matx="190_"$(basename "$tmp_matx")".tsv"
+    
+      printf "${bold}$(date):${normal} Exporting \"$(basename "${inpth_pcoa[$i]}")\".\n"
+      # erase possibly existing temp files
+      rm -f "$TMPDIR"ordination.txt
+      # export to temp file
+      qiime tools export \
+        --input-path  "${inpth_pcoa[$i]}" \
+        --output-path "$TMPDIR"
+      # move temp file in place
+      mv "$TMPDIR"ordination.txt "$results_dir"/"$results_pcoa"
+
+      printf "${bold}$(date):${normal} Exporting \"$(basename "${inpth_matrix[$i]}")\".\n"
+      # erase possibly existing temp files
+      rm -f "$TMPDIR"distance-matrix.tsv 
+      # export to temp file
+      qiime tools export \
+        --input-path  "${inpth_matrix[$i]}" \
+        --output-path "$TMPDIR"
+      # move temp file in place
+      mv "$TMPDIR"distance-matrix.tsv "$results_dir"/"$results_matx"
+   
+   else
+
+    # diagnostic message
+    printf "${bold}$(date):${normal} Detected readily available results, skipping export of one file set.\n"
+
+  fi
 
   else
